% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zot_db.R
\name{zot_creators_raw}
\alias{zot_creators_raw}
\title{Zotero Mock Creators Data}
\format{
A data frame (tibble) with 20 rows and 4 variables:
\describe{
\item{creatorID}{Primary key for the creator.}
\item{lastName}{Last name or institutional name.}
\item{firstName}{First name.}
\item{fieldMode}{0 for two-field mode (persons), 1 for single-field mode (institutions).}
}
}
\usage{
zot_creators_raw
}
\description{
A dataset simulating the 'creators' table in a Zotero database.
Contains messy author names, capitalization errors, swapped Asian names,
and misclassified institutions to test cleaning functions.
}
