% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zot_db.R
\name{zot_disconnect_db}
\alias{zot_disconnect_db}
\title{Disconnect from the Zotero Database}
\usage{
zot_disconnect_db(con)
}
\arguments{
\item{con}{An active DBI connection.}
}
\value{
Invisible TRUE if successful.
}
\description{
Safely closes the connection to the SQLite database (mock or real) and
provides a status message.
}
