% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zot_authors.R
\name{zot_find_authors}
\alias{zot_find_authors}
\title{Find Authors in the Zotero Database}
\usage{
zot_find_authors(con, pattern, ignore_case = TRUE)
}
\arguments{
\item{con}{An active DBI connection to a Zotero database.}

\item{pattern}{A character string containing a regular expression to search for.}

\item{ignore_case}{Logical. Should the search be case-insensitive? Default is TRUE.}
}
\value{
A tibble containing the matching creators.
}
\description{
Searches the 'creators' table for a specific pattern in either the first
or last name. This is the first step before merging duplicate authors.
}
