% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zot_publishers.R
\name{zot_find_publishers}
\alias{zot_find_publishers}
\title{Find Publishers in the Zotero Database}
\usage{
zot_find_publishers(con, pattern, ignore_case = TRUE)
}
\arguments{
\item{con}{An active DBI connection to a Zotero database.}

\item{pattern}{A character string containing a regular expression to search for.}

\item{ignore_case}{Logical. Should the search be case-insensitive? Default is TRUE.}
}
\value{
A tibble containing the matching publisher strings and their valueIDs.
}
\description{
Searches the 'itemDataValues' table for specific publisher names.
This is the first step before merging duplicate or inconsistently named publishers.
}
