% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{zot_item_creators_raw}
\alias{zot_item_creators_raw}
\title{Zotero Mock Item Creators Link Data}
\format{
A data frame (tibble) with 15 rows and 4 variables:
\describe{
\item{itemID}{Foreign key referencing items.}
\item{creatorID}{Foreign key referencing creators.}
\item{creatorTypeID}{Type of creator (usually 1 for Author).}
\item{orderIndex}{Position in the author list (0 = first author).}
}
}
\usage{
zot_item_creators_raw
}
\description{
A dataset simulating the 'itemCreators' table, linking publications (items)
to their authors (creators) and defining author order.
}
\keyword{datasets}
