% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zot_db.R
\name{zot_item_data_raw}
\alias{zot_item_data_raw}
\title{Zotero Mock Item Data Link Table}
\format{
A data frame (tibble) with 32 rows and 3 variables:
\describe{
\item{itemID}{Foreign key referencing items.}
\item{fieldID}{Foreign key referencing fields.}
\item{valueID}{Foreign key referencing itemDataValues.}
}
}
\usage{
zot_item_data_raw
}
\description{
A dataset simulating the 'itemData' table, linking publications to their
specific metadata values.
}
