% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{zot_item_data_values_raw}
\alias{zot_item_data_values_raw}
\title{Zotero Mock Item Data Values}
\format{
A data frame (tibble) with 32 rows and 2 variables:
\describe{
\item{valueID}{Primary key for the string value.}
\item{value}{The actual text string.}
}
}
\usage{
zot_item_data_values_raw
}
\description{
A dataset simulating the 'itemDataValues' table, storing all unique
metadata strings (titles, journal names, dates, etc.).
}
\keyword{datasets}
