% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{zot_items_raw}
\alias{zot_items_raw}
\title{Zotero Mock Items Data}
\format{
A data frame (tibble) with 15 rows and 2 variables:
\describe{
\item{itemID}{Primary key for the item (publication).}
\item{itemTypeID}{Type of publication (1 = Journal Article, 2 = Book, 3 = Conference Paper).}
}
}
\usage{
zot_items_raw
}
\description{
A dataset simulating the 'items' table in a Zotero database.
}
\keyword{datasets}
