% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zot_db.R
\name{zot_mock_db}
\alias{zot_mock_db}
\title{Create a Realistic In-Memory Zotero Test Database}
\usage{
zot_mock_db()
}
\value{
A DBI connection to a temporary, populated SQLite database in RAM.
}
\description{
This function simulates Zotero's exact relational database schema in-memory.
It utilizes pre-defined package datasets (e.g., zot_creators_raw) to populate
the SQLite tables.
}
